NTP-640: refactor code to move records/classes, enums and extensions in to own files and folders
