+++ Domain/TuitionType.cs
@@ -13,7 +13,7 @@ public class TuitionType
 
     public override bool Equals(object? obj)
     {
-        if (ReferenceEquals(null, obj)) return false;
+        if (obj is null) return false;
         if (ReferenceEquals(this, obj)) return true;
         if (obj.GetType() != this.GetType()) return false;
         return Equals((TuitionType)obj);
