+++ UI/Enums/ReferrerList.cs
@@ -1,0 +1,8 @@
+﻿namespace UI.Enums
+{
+    public enum ReferrerList
+    {
+        SearchResults = 0,
+        FullList = 1
+    }
+}
