+++ UI/Enums/Subject.cs
@@ -1,0 +1,14 @@
+﻿namespace UI.Enums
+{
+    public enum Subject
+    {
+        Unspecified = 0,
+
+        English,
+        Maths,
+        Science,
+
+        Humanities,
+        ModernForeignLanguages,
+    }
+}
