+++ UI/Extensions/KeyStageSubjectExtensions.cs
@@ -1,0 +1,12 @@
+﻿using UI.Models;
+
+namespace UI.Extensions
+{
+    public static class KeyStageSubjectExtensions
+    {
+        public static KeyStageSubject[] ParseKeyStageSubjects(this string[] keyStageSubjects)
+            => keyStageSubjects.Select(KeyStageSubject.TryParse)
+            .OfType<KeyStageSubject>()
+            .ToArray();
+    }
+}
