+++ UI/Pages/Index.cshtml.cs
@@ -5,6 +5,7 @@ using FluentValidation;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
+using UI.Models;
 
 namespace UI.Pages;
