--- UI/Pages/Selectable.cs
@@ -1,9 +1,0 @@
-﻿namespace UI.Pages;
-
-public class Selectable : Selectable<string> { }
-
-public class Selectable<T>
-{
-    public T Name { get; set; } = default!;
-    public bool Selected { get; set; }
-}
