+++ UI/Pages/TuitionPartner.cshtml.cs
@@ -6,6 +6,7 @@ using MediatR;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.EntityFrameworkCore;
+using UI.Models;
 
 namespace UI.Pages;
 
@@ -87,7 +88,7 @@ public class TuitionPartner : PageModel
         string[] TuitionTypes, string[] Ratios, Dictionary<int, GroupPrice> Prices,
         string Website, string PhoneNumber, string EmailAddress, string[] Address, bool HasSenProvision, bool IsVatCharged,
         LocalAuthorityDistrictCoverage[] LocalAuthorityDistricts,
-        Dictionary<TuitionType, Dictionary<KeyStage, Dictionary<string, Dictionary<int, decimal>>>> AllPrices, string LegalStatus)
+        Dictionary<Enums.TuitionType, Dictionary<Enums.KeyStage, Dictionary<string, Dictionary<int, decimal>>>> AllPrices, string LegalStatus)
     {
         public bool HasPricingVariation => Prices.Any(x => x.Value.HasVariation);
     }
@@ -140,7 +141,7 @@ public class TuitionPartner : PageModel
 
             if (tp == null) return null;
 
-            var subjects = tp.SubjectCoverage.Select(x => x.Subject).Distinct().GroupBy(x => x.KeyStageId).Select(x => $"{((KeyStage)x.Key).DisplayName()} - {x.DisplayList()}");
+            var subjects = tp.SubjectCoverage.Select(x => x.Subject).Distinct().GroupBy(x => x.KeyStageId).Select(x => $"{((Enums.KeyStage)x.Key).DisplayName()} - {x.DisplayList()}");
             string? ladCode = null;
             if (!string.IsNullOrEmpty(request.Postcode))
             {
@@ -148,7 +149,7 @@ public class TuitionPartner : PageModel
                 ladCode = location?.LocalAuthorityDistrictCode;
             }
             var types = await GetTuitionTypesCovered(ladCode, tp, cancellationToken);
-            if (types == null || types.Count() == 0)
+            if (types == null || !types.Any())
             {
                 //If no data returned then postcode is invalid, has been changed so does not apply for the TP
                 //or issue calling GetLocationFilterParametersAsync (calling postcode.io)
@@ -212,7 +213,7 @@ public class TuitionPartner : PageModel
 
             var coverageDictionary = coverage
                 .GroupBy(e => e.TuitionTypeId)
-                .ToDictionary(e => (TuitionType)e.Key, e => e.ToDictionary(x => x.LocalAuthorityDistrictId, x => x));
+                .ToDictionary(e => (Enums.TuitionType)e.Key, e => e.ToDictionary(x => x.LocalAuthorityDistrictId, x => x));
 
             var regions = await _db.Regions
                 .Include(e => e.LocalAuthorityDistricts.OrderBy(x => x.Code))
@@ -225,8 +226,8 @@ public class TuitionPartner : PageModel
             {
                 foreach (var lad in lads)
                 {
-                    var inSchool = coverageDictionary.ContainsKey(TuitionType.InSchool) && coverageDictionary[TuitionType.InSchool].ContainsKey(lad.Id);
-                    var online = coverageDictionary.ContainsKey(TuitionType.Online) && coverageDictionary[TuitionType.Online].ContainsKey(lad.Id);
+                    var inSchool = coverageDictionary.ContainsKey(Enums.TuitionType.InSchool) && coverageDictionary[Enums.TuitionType.InSchool].ContainsKey(lad.Id);
+                    var online = coverageDictionary.ContainsKey(Enums.TuitionType.Online) && coverageDictionary[Enums.TuitionType.Online].ContainsKey(lad.Id);
                     result[lad.Id] = new LocalAuthorityDistrictCoverage(region.Name, lad.Code, lad.Name, inSchool, online);
                 }
             }
@@ -276,16 +277,16 @@ public class TuitionPartner : PageModel
             }
         }
 
-        private async Task<Dictionary<TuitionType, Dictionary<KeyStage, Dictionary<string, Dictionary<int, decimal>>>>> GetFullPricing(Query request, ICollection<Price> prices)
+        private async Task<Dictionary<Enums.TuitionType, Dictionary<Enums.KeyStage, Dictionary<string, Dictionary<int, decimal>>>>> GetFullPricing(Query request, ICollection<Price> prices)
         {
             if (!request.ShowFullPricing) return new();
 
-            var fullPricing = new Dictionary<TuitionType, Dictionary<KeyStage, Dictionary<string, Dictionary<int, decimal>>>>();
+            var fullPricing = new Dictionary<Enums.TuitionType, Dictionary<Enums.KeyStage, Dictionary<string, Dictionary<int, decimal>>>>();
 
-            foreach (var tuitionType in new[] { TuitionType.InSchool, TuitionType.Online })
+            foreach (var tuitionType in new[] { Enums.TuitionType.InSchool, Enums.TuitionType.Online })
             {
-                fullPricing[tuitionType] = new Dictionary<KeyStage, Dictionary<string, Dictionary<int, decimal>>>();
-                foreach (var keyStage in new[] { KeyStage.KeyStage1, KeyStage.KeyStage2, KeyStage.KeyStage3, KeyStage.KeyStage4 })
+                fullPricing[tuitionType] = new Dictionary<Enums.KeyStage, Dictionary<string, Dictionary<int, decimal>>>();
+                foreach (var keyStage in new[] { Enums.KeyStage.KeyStage1, Enums.KeyStage.KeyStage2, Enums.KeyStage.KeyStage3, Enums.KeyStage.KeyStage4 })
                 {
                     fullPricing[tuitionType][keyStage] = new Dictionary<string, Dictionary<int, decimal>>();
 
@@ -299,8 +300,8 @@ public class TuitionPartner : PageModel
 
             foreach (var price in prices)
             {
-                var tuitionType = (TuitionType)price.TuitionTypeId;
-                var keyStage = (KeyStage)price.Subject.KeyStageId;
+                var tuitionType = (Enums.TuitionType)price.TuitionTypeId;
+                var keyStage = (Enums.KeyStage)price.Subject.KeyStageId;
                 var subjectName = price.Subject.Name;
 
                 fullPricing[tuitionType][keyStage][subjectName][price.GroupSize] = price.HourlyRate;
