+++ UI/Pages/WhichKeyStages.cshtml.cs
@@ -2,6 +2,8 @@ using FluentValidation;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
+using UI.Enums;
+using UI.Models;
 
 namespace UI.Pages
 {
