+++ UI/Pages/WhichSubjects.cshtml.cs
@@ -2,6 +2,9 @@ using FluentValidation;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
+using UI.Enums;
+using UI.Extensions;
+using UI.Models;
 
 namespace UI.Pages;
