+++ UI/Pages/_ViewImports.cshtml
@@ -1,4 +1,5 @@
 ﻿@using UI.Models
+@using UI.Enums
 @using UI.Pages.Shared
 @using Application.Extensions
 @using UI.Extensions
